+++ backend/tpbackend/api_v2_models.py
@@ -15,15 +15,6 @@ class PublicPlatformModelV2(BaseModel):
     updated: int
 
 
-class PublicUserModelV2(BaseModel):
-    id: int
-    discord_id: str | None
-    name: str
-    default_platform_id: int
-    created: int
-    updated: int
-
-
 class PublicGameModelV2(BaseModel):
     id: int
     name: str
@@ -56,25 +47,3 @@ class GameStatsV2(PublicGameModelV2, GameTotals):
 
 class PlatformStatsV2(PublicPlatformModelV2, PlatformTotals):
     pass
-
-
-class UserStatsV2(PublicUserModelV2):
-    stats: UserTotals
-
-    @classmethod
-    def from_user(cls, user):
-        return cls(
-            id=user.id,
-            discord_id=user.discord_id,
-            name=user.name,
-            default_platform_id=user.default_platform_id,
-            created=dt_to_ts(user.created),
-            updated=dt_to_ts(user.updated),
-            stats=UserTotals(
-                seconds=user.total_seconds,
-                activity_count=user.activity_count,
-                game_count=user.game_count,
-                platform_count=user.platform_count,
-                last_activity=dt_to_ts(user.last_activity),
-            ),
-        )
